Is Equal mover to HtmSerializer2
